adding some background to get opinion miner results displayed.

+++ sare-abome/src/main/java/eu/ubipol/opinionmining/nlp_engine/OpinionWords.java
@@ -28,16 +28,16 @@ import eu.ubipol.opinionmining.nlp_engine.Utils.WordType;
 
 
 public class OpinionWords {
-	private static Map<String, Float> wordScores;
+	private static Map<String, Double> wordScores;
 
 	public static void fillScoreList(File file) {
 		Scanner reader = null;
 		try {
-			wordScores = new HashMap<String, Float>();
+			wordScores = new HashMap<String, Double>();
 			reader = new Scanner(file);
 			while (reader.hasNextLine()) {
 				String[] line = reader.nextLine().split("\t");
-				wordScores.put(line[0], Float.valueOf(line[1]));
+				wordScores.put(line[0], Double.valueOf(line[1]));
 			}
 		} catch (Exception e) {
 		} finally {
@@ -50,13 +50,13 @@ public class OpinionWords {
 	private static void fillScoreList() {
 		Scanner reader = null;
 		try {
-			wordScores = new HashMap<String, Float>();
+			wordScores = new HashMap<String, Double>();
 			reader = new Scanner(
 					OpinionWords.class
 							.getResourceAsStream("/sentiwordnet_stemmed.txt"));
 			while (reader.hasNextLine()) {
 				String[] line = reader.nextLine().split("\t");
-				wordScores.put(line[0], Float.valueOf(line[1]));
+				wordScores.put(line[0], Double.valueOf(line[1]));
 			}
 		} catch (Exception e) {
 		} finally {
@@ -66,7 +66,7 @@ public class OpinionWords {
 		}
 	}
 
-	private static Map<String, Float> getScoreList() {
+	private static Map<String, Double> getScoreList() {
 		try {
 			if (wordScores == null) {
 				fillScoreList();
@@ -78,7 +78,7 @@ public class OpinionWords {
 		}
 	}
 
-	protected static Float getWordScore(String word, WordType wordType) {
+	protected static Double getWordScore(String word, WordType wordType) {
 		String searchString = word;
 		if (wordType == WordType.ADJECTIVE)
 			searchString += "_a";
@@ -89,10 +89,10 @@ public class OpinionWords {
 		else if (wordType == WordType.NOUN)
 			searchString += "_n";
 		else
-			return new Float(-2);
+			return new Double(-2);
 		if (getScoreList().containsKey(searchString))
 			return getScoreList().get(searchString);
 		else
-			return new Float(-2);
+			return new Double(-2);
 	}
 }
